+++ recipes/linux_standalone.rb
@@ -51,29 +51,14 @@ execute 'Generating Self-Signed Java Keystore' do
   only_if { settings['tomcat']['keystoreFile'] == "#{node['confluence']['home_path']}/.keystore" }
 end
 
-remote_file "#{Chef::Config[:file_cache_path]}/atlassian-confluence-#{node['confluence']['version']}.tar.gz" do
-  source node['confluence']['url']
+ark 'confluence' do
+  url node['confluence']['url']
+  prefix_root File.dirname(node['confluence']['install_path'])
+  home_dir node['confluence']['install_path']
   checksum node['confluence']['checksum']
-  mode '0644'
-  action :create
-end
-
-directory File.dirname(node['confluence']['install_path']) do
+  version node['confluence']['version']
   owner node['confluence']['user']
   group node['confluence']['user']
-  mode 00755
-  action :create
-  recursive true
-end
-
-execute "Extracting Confluence #{node['confluence']['version']}" do
-  cwd Chef::Config[:file_cache_path]
-  command <<-COMMAND
-    tar -zxf atlassian-confluence-#{node['confluence']['version']}.tar.gz
-    mv atlassian-confluence-#{node['confluence']['version']} #{node['confluence']['install_path']}
-    chown -R #{node['confluence']['user']} #{node['confluence']['install_path']}
-  COMMAND
-  creates "#{node['confluence']['install_path']}/confluence"
 end
 
 if settings['database']['type'] == 'mysql'